backend/apps/clients/serializers.py
from rest_framework import serializers
from .models import Client


class ClientSerializer(serializers.ModelSerializer):
    class Meta:
        model = Client
        fields = ('id', 'client_id', 'portfolio_market_value', 'portfolio_profit')
