backend/requirements.txt
# postgres database adapater
psycopg2==2.6.2

# django framework itself
Django==1.11.5

#app server for prod
gunicorn==19.6.0

#selenium for e2e tests
selenium==3.0.2

# Django Rest Framework (DRF)
djangorestframework==3.6.4

# pyexcel_xlsx
pyexcel-xlsx==0.5.0.1
# pyexcel_xlsx [dependencies]
pyexcel-io==0.5.1
openpyxl==2.4.8
lml==0.0.1
jdcal==1.3
et-xmlfile==1.0.1

# django extensions
django-extensions==1.9.1
# django extensions [dependencies]
six==1.10.0  # also a dependency of factory_boy

# factory_boy
factory-boy==2.9.2
# factory_boy [dependencies]
Faker==0.8.3
dnspython==1.15.0
email-validator==1.0.2
idna==2.6
python-dateutil==2.6.1

